Master_26_06_2018

+++ app/Http/Controllers/API/PurchaseRequestDetailsAPIController.php
@@ -18,15 +18,19 @@ use App\Models\CompanyPolicyMaster;
 use App\Models\FinanceItemcategorySubAssigned;
 use App\Models\GRVDetails;
 use App\Models\ItemAssigned;
+use App\Models\ProcumentOrder;
 use App\Models\PurchaseOrderDetails;
 use App\Models\PurchaseRequest;
 use App\Models\PurchaseRequestDetails;
 use App\Repositories\PurchaseRequestDetailsRepository;
+use App\Repositories\PurchaseRequestRepository;
 use Illuminate\Http\Request;
 use App\Http\Controllers\AppBaseController;
 use InfyOm\Generator\Criteria\LimitOffsetCriteria;
 use Prettus\Repository\Criteria\RequestCriteria;
 use Response;
+use Illuminate\Support\Facades\DB;
+
 
 /**
  * Class PurchaseRequestDetailsController
@@ -37,9 +41,12 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
     /** @var  PurchaseRequestDetailsRepository */
     private $purchaseRequestDetailsRepository;
 
-    public function __construct(PurchaseRequestDetailsRepository $purchaseRequestDetailsRepo)
+    private $purchaseRequestRepository;
+
+    public function __construct(PurchaseRequestDetailsRepository $purchaseRequestDetailsRepo,PurchaseRequestRepository $purchaseRequestRepo)
     {
         $this->purchaseRequestDetailsRepository = $purchaseRequestDetailsRepo;
+        $this->purchaseRequestRepository = $purchaseRequestRepo;
     }
 
     /**
@@ -91,9 +98,10 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
         $input = array_except($request->all(), 'uom');
         $input = $this->convertArrayToValue($input);
 
+
         $companySystemID = $input['companySystemID'];
 
-        $item = ItemAssigned::where('itemCodeSystem', $input['itemCode'])
+         $item = ItemAssigned::where('itemCodeSystem', $input['itemCode'])
             ->where('companySystemID', $companySystemID)
             ->first();
 
@@ -102,15 +110,22 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
         }
 
         $purchaseRequest = PurchaseRequest::where('purchaseRequestID', $input['purchaseRequestID'])
-            ->first();
-
-        $input['budgetYear'] = $purchaseRequest->budgetYear;
+                                           ->first();
 
 
         if (empty($purchaseRequest)) {
             return $this->sendError('Purchase Request Details not found');
         }
 
+        if($purchaseRequest->cancelledYN == -1){
+            return $this->sendError('This Purchase Request already closed. You can not add.',500);
+        }
+
+        if($purchaseRequest->approved == 1){
+            return $this->sendError('This Purchase Request fully approved. You can not add.',500);
+        }
+
+        $input['budgetYear'] = $purchaseRequest->budgetYear;
         $input['itemPrimaryCode'] = $item->itemPrimaryCode;
         $input['itemDescription'] = $item->itemDescription;
         $input['partNumber'] = $item->secondaryItemCode;
@@ -118,7 +133,7 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
         $input['itemFinanceCategorySubID'] = $item->financeCategorySub;
         //$input['estimatedCost'] = $item->wacValueLocal;
 
-        $currencyConversion = \Helper::currencyConversion($item->companySystemID,$item->wacValueLocalCurrencyID,$purchaseRequest->currency, $item->wacValueLocal);
+        $currencyConversion = \Helper::currencyConversion($item->companySystemID, $item->wacValueLocalCurrencyID, $purchaseRequest->currency, $item->wacValueLocal);
 
         $input['estimatedCost'] = $currencyConversion['documentAmount'];
 
@@ -134,7 +149,7 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
             ->first();
 
         if (empty($financeItemCategorySubAssigned)) {
-            return $this->sendError('Item not found');
+            return $this->sendError('Finance Category not found');
         }
 
 
@@ -157,6 +172,9 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
         $input['financeGLcodePL'] = $financeItemCategorySubAssigned->financeGLcodePL;
         $input['includePLForGRVYN'] = $financeItemCategorySubAssigned->includePLForGRVYN;
 
+        $input['itemCategoryID'] = 0;
+        $input['maxQty'] = 0;
+        $input['minQty'] = 0;
 
         // check policy 18
 
@@ -192,20 +210,20 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
 
                 $anyPendingApproval = $checkWhether->whereHas('details', function ($query) use ($companySystemID, $purchaseRequest, $item) {
                     $query->where('itemPrimaryCode', $item->itemPrimaryCode);
-                    /* $query->groupBy(
-                         'erp_purchaserequestdetails.itemCode',
-                         'erp_purchaserequestdetails.itemPrimaryCode',
-                         'erp_purchaserequestdetails.selectedForPO',
-                         'erp_purchaserequestdetails.prClosedYN',
-                         'erp_purchaserequestdetails.fullyOrdered'
-                     )->select([
+                                /* $query->groupBy(
+                                     'erp_purchaserequestdetails.itemCode',
+                                     'erp_purchaserequestdetails.itemPrimaryCode',
+                                     'erp_purchaserequestdetails.selectedForPO',
+                                     'erp_purchaserequestdetails.prClosedYN',
+                                     'erp_purchaserequestdetails.fullyOrdered'
+                                 )->select([
                                  'erp_purchaserequestdetails.itemCode',
                                  'erp_purchaserequestdetails.itemPrimaryCode',
                                  'erp_purchaserequestdetails.selectedForPO',
                                  'erp_purchaserequestdetails.prClosedYN',
                                  'erp_purchaserequestdetails.fullyOrdered'
                               ]);*/
-                })
+                    })
                     ->where('approved', 0)
                     ->where('cancelledYN', 0)
                     ->first();
@@ -214,7 +232,6 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
                 if (!empty($anyPendingApproval)) {
                     return $this->sendError("There is a purchase request (" . $anyPendingApproval->purchaseRequestCode . ") pending for approval for the item you are trying to add. Please check again.", 500);
                 }
-
 
                 $anyApprovedPRButPONotProcessed = PurchaseRequest::where('purchaseRequestID', '!=', $purchaseRequest->purchaseRequestID)
                     ->where('companySystemID', $companySystemID)
@@ -242,13 +259,13 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
                             ->where('selectedForPO', 0)
                             ->where('prClosedYN', 0)
                             ->where('fullyOrdered', 0);
-                        /* $query->groupBy(
-                             'erp_purchaserequestdetails.itemCode',
-                             'erp_purchaserequestdetails.itemPrimaryCode',
-                             'erp_purchaserequestdetails.selectedForPO',
-                             'erp_purchaserequestdetails.prClosedYN',
-                             'erp_purchaserequestdetails.fullyOrdered'
-                         )->select([
+                                    /* $query->groupBy(
+                                         'erp_purchaserequestdetails.itemCode',
+                                         'erp_purchaserequestdetails.itemPrimaryCode',
+                                         'erp_purchaserequestdetails.selectedForPO',
+                                         'erp_purchaserequestdetails.prClosedYN',
+                                         'erp_purchaserequestdetails.fullyOrdered'
+                                     )->select([
                                      'erp_purchaserequestdetails.itemCode',
                                      'erp_purchaserequestdetails.itemPrimaryCode',
                                      'erp_purchaserequestdetails.selectedForPO',
@@ -290,13 +307,13 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
                             ->where('selectedForPO', 0)
                             ->where('prClosedYN', 0)
                             ->where('fullyOrdered', 1);
-                        /* $query->groupBy(
-                             'erp_purchaserequestdetails.itemCode',
-                             'erp_purchaserequestdetails.itemPrimaryCode',
-                             'erp_purchaserequestdetails.selectedForPO',
-                             'erp_purchaserequestdetails.prClosedYN',
-                             'erp_purchaserequestdetails.fullyOrdered'
-                         )->select([
+                                    /* $query->groupBy(
+                                         'erp_purchaserequestdetails.itemCode',
+                                         'erp_purchaserequestdetails.itemPrimaryCode',
+                                         'erp_purchaserequestdetails.selectedForPO',
+                                         'erp_purchaserequestdetails.prClosedYN',
+                                         'erp_purchaserequestdetails.fullyOrdered'
+                                     )->select([
                                      'erp_purchaserequestdetails.itemCode',
                                      'erp_purchaserequestdetails.itemPrimaryCode',
                                      'erp_purchaserequestdetails.selectedForPO',
@@ -313,7 +330,22 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
                     return $this->sendError("There is a purchase request (" . $anyApprovedPRButPOPartiallyProcessed->purchaseRequestCode . ") approved and PO is partially processed for the item you are trying to add. Please check again", 500);
                 }
 
-                /* $unApprovedPO*/
+                /* PO check*/
+
+                $checkPOPending  = ProcumentOrder::where('companySystemID', $companySystemID)
+                                                    ->where('serviceLineSystemID', $purchaseRequest->serviceLineSystemID)
+                                                    ->whereHas('detail', function ($query) use ($item) {
+                                                        $query->where('itemPrimaryCode', $item->itemPrimaryCode);
+                                                    })
+                                                    ->where('approved', 0)
+                                                    ->where('poCancelledYN', 0)
+                                                    ->first();
+
+                if (!empty($checkPOPending)) {
+                    return $this->sendError("There is a purchase order (" . $checkPOPending->purchaseOrderCode . ") pending for approval for the item you are trying to add. Please check again.", 500);
+                }
+                /* PO --> approved=-1 And cancelledYN=0 */
+
             }
         }
 
@@ -353,11 +385,11 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
             ->sum('noQty');
 
         $quantityOnOrder = $poQty - $grvQty;
-        $quantityInHand = $poQty;
+        $quantityInHand  = $poQty;
 
-        $input['poQuantity'] = $poQty;
+        $input['poQuantity']      = $poQty;
         $input['quantityOnOrder'] = $quantityOnOrder;
-        $input['quantityInHand'] = $quantityInHand;
+        $input['quantityInHand']  = $quantityInHand;
 
         $purchaseRequestDetails = $this->purchaseRequestDetailsRepository->create($input);
 
@@ -405,6 +437,16 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
         if (empty($purchaseRequestDetails)) {
             return $this->sendError('Purchase Request Details not found');
         }
+        $purchaseRequest = $this->purchaseRequestRepository->findWithoutFail($input['purchaseRequestID']);
+        if (empty($purchaseRequest)) {
+            return $this->sendError('Purchase Request not found');
+        }
+        if($purchaseRequest->cancelledYN == -1){
+            return $this->sendError('This Purchase Request already closed. You can not edit.',500);
+        }
+        if($purchaseRequest->approved == 1){
+            return $this->sendError('This Purchase Request fully approved. You can not edit.',500);
+        }
 
         $purchaseRequestDetails = $this->purchaseRequestDetailsRepository->update($input, $id);
 
@@ -428,6 +470,17 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
             return $this->sendError('Purchase Request Details not found');
         }
 
+        $purchaseRequest = $this->purchaseRequestRepository->findWithoutFail($purchaseRequestDetails->purchaseRequestID);
+        if (empty($purchaseRequest)) {
+            return $this->sendError('Purchase Request not found');
+        }
+        if($purchaseRequest->cancelledYN == -1){
+            return $this->sendError('This Purchase Request already closed. You can not delete.',500);
+        }
+        if($purchaseRequest->approved == 1){
+            return $this->sendError('This Purchase Request fully approved. You can not delete.',500);
+        }
+
         $purchaseRequestDetails->delete();
 
         return $this->sendResponse($id, 'Purchase Request Details deleted successfully');
@@ -444,15 +497,11 @@ class PurchaseRequestDetailsAPIController extends AppBaseController
     public function getPurchaseRequestDetailForPO(Request $request)
     {
         $input = $request->all();
-        $prId = $input['purchaseRequestID'];
+        $prID = $input['purchaseRequestID'];
 
-        $detail = PurchaseRequestDetails::select(DB::raw('erp_purchaserequestdetails.*,"" as isChecked, "" as poQty'))
-            ->where('purchaseRequestID', $prId)
-            //->where('selectedForPO', 0)
-            ->where('prClosedYN', 0)
-            ->where('fullyOrdered','!=', 2)
-            ->get();
-        return $this->sendResponse($detail->toArray(), 'Purchase Request Details retrieved successfully');
+        $detail = DB::select('SELECT prdetails.*,"" as isChecked, "" as poQty, 0 as poUnitAmount,podetails.poTakenQty FROM erp_purchaserequestdetails prdetails LEFT JOIN (SELECT erp_purchaseorderdetails.purchaseRequestDetailsID, SUM(noQty) AS poTakenQty FROM erp_purchaseorderdetails GROUP BY purchaseRequestDetailsID,itemCode) as podetails ON prdetails.purchaseRequestDetailsID = podetails.purchaseRequestDetailsID WHERE purchaseRequestID = '.$prID.' AND prClosedYN = 0 AND fullyOrdered != 2 AND manuallyClosed = 0');
+
+        return $this->sendResponse($detail, 'Purchase Request Details retrieved successfully');
 
     }
 }
